+++ examples/contracts/src/lib.rs
@@ -3,6 +3,7 @@ mod tests {
     use std::{collections::HashSet, time::Duration};
 
     use fuels::{
+        accounts::signers::{fake::FakeSigner, private_key::PrivateKeySigner},
         core::codec::{encode_fn_selector, ABIFormatter, DecoderConfig, EncoderConfig},
         prelude::{LoadConfiguration, NodeConfig, StorageConfiguration},
         programs::debug::ScriptType,
